Added suspend, ban, & reactivate functionality; Includes modals, minor cleanup and extra detail
Buttons to Suspend, Ban, or Reactivate an account were added to change an account's status. Some appear or disappear depending on said status. Every button is connected to a confirmation modal.
For Suspend, you can decide on a suspension period ranging from 6 hours to 7 days. The date that the suspension ends (on its own) will appear on the table. If user status is changed again, it will disappear.
The format of the Users table was edited slightly to fit the buttons in one line, and the Log Out button now looks like the rest of the buttons.